Avoid replay added/remove issues.
This is done by listening to the 'sync' game in multiplayer instead.
This was we don't have to do anything during replay. The drawback is
that the removal of the bodies will be slightly delayed. But on the
plus-side the visuals will be completely in sync with the game state.