Rewrote tick to use action names instead of callbacks.
The reason is because copying of callbacks between game and sync seems
to work badly in multiplayer. So the idea is that only using primitives
and pass around a string (action name) plus an array of ids and booleans
would work better.

This also has the benefit of getting the extra logic into their own
actions.