Worked on multiplayer time-shifting.
By changing the speed of the puck we gain a few frames to be able to
move the opponents paddle to the position where it will hit. And after
it's been visually hit we run a replay.

Also made it move the opponent paddle as soon as it it received for a
much more alive experience.

TODO
- Adjust the puck speed dynamically based on latency and velocity.
- Use body.target to smooth out the movements of the opponents paddle.
  Or maybe a simple 'easing' might be enough. As long as the 'hit' event
will move it to the exact position.